reorganize, del WIP files, comment out mains
