+++ src/search/algorithm.hh
@@ -2,10 +2,14 @@
 
 #include "../tree/node.hh"
 
+#include <thread>
+#include <mutex>
+
 template <class _Model>
 class AbstractAlgorithm
 {
-    static_assert(std::derived_from<_Model, DualPolicyValueModel<typename _Model::Types::State>>);
+    // static_assert(std::derived_from<_Model, DualPolicyValueModel<typename _Model::Types::State>>,
+    // "Even AbstractAlgorithm currently requires a DualPolicyValueModel");
 
 public:
     struct MatrixStats
@@ -22,22 +26,16 @@ public:
     };
 };
 
-/*
-Node local bandit algorithm.
-*/
-
-template <class Model>
-class BanditAlgorithm : public AbstractAlgorithm<Model>
+template <class Model, class Algorithm>
+class TreeBandit : public AbstractAlgorithm<Model>
 {
-    static_assert(std::derived_from<Model, AbstractModel<typename Model::Types::State>>);
-    // We don't assume that Model is a DVPmodel.
-
 public:
-    struct Outcome; // If you don't add to types, you have to typename BanditAlgorithm<Model> in the specification...
+    struct Outcome;
     struct Types : AbstractAlgorithm<Model>::Types
     {
-        using Outcome = BanditAlgorithm::Outcome;
+        using Outcome = TreeBandit::Outcome;
     };
+
     struct Outcome
     {
         int row_idx, col_idx;
@@ -45,70 +43,22 @@ public:
         typename Types::Real row_mu, col_mu; // Actor policy at row_idx, col_idx.
     };
 
-    void select(
-        MatrixNode<BanditAlgorithm> *matrix_node,
-        Outcome &outcome);
-    void expand(
-        typename Types::State &state,
-        typename Types::Model &model,
-        MatrixNode<BanditAlgorithm> *matrix_node);
-    void init_stats(
-        typename Types::State &state,
-        typename Types::Model &model,
-        MatrixNode<BanditAlgorithm> *matrix_node);
-    void update_matrix_node(
-        MatrixNode<BanditAlgorithm> *matrix_node,
-        Outcome &outcome);
-    void update_chance_node(
-        ChanceNode<BanditAlgorithm> *chance_node,
-        Outcome &outcome);
-};
-
-/*
-This algorithm generalizes the typical "MCTS" pattern found in UCT, PUCT, SM-MCTS(-A), MatrixUCB etc.
-This will also be subclassed for multithreaded implementations.
-*/
-
-template <class Algorithm> // TODO consider make template params reflect "LowerDerived"
-class TreeBandit : public Algorithm
-{
-    static_assert(std::derived_from<Algorithm, BanditAlgorithm<typename Algorithm::Types::Model>>,
-        "TreeBandit only accepts a BanditAlgorithm as its template paramater"
-    );
-
-public:
-    struct MatrixStats;
-    struct ChanceStats;
-    struct Types : Algorithm::Types
-    {
-        using MatrixStats = TreeBandit::MatrixStats;
-        using ChanceStats = TreeBandit::ChanceStats;
-    };
-
-    // TreeBandit(prng &device) : Algorithm(device) {}
-
-    using Algorithm::Algorithm;
-    // Constructor forwarding. I consider this take care of for now.
-
-    // template <typename ...Args>
-    // TreeBandit (Args... args) : Algorithm(args...) {}
-    // Future me: If you use template (...Args) type stuff then the prng device at the bandit level will not perform sample_pdf right. It's fucked idk
-
     void run(
         int playouts,
         typename Types::State &state,
         typename Types::Model &model,
         MatrixNode<Algorithm> &matrix_node)
     {
-        Algorithm::init_stats(playouts, state, model, &matrix_node);
+        init_stats(playouts, state, model, &matrix_node);
         for (int playout = 0; playout < playouts; ++playout)
         {
-            _playout(state, model, &matrix_node);
+            this->playout(state, model, &matrix_node);
         }
     }
 
 private:
-    MatrixNode<Algorithm> *_playout(
+
+    MatrixNode<Algorithm> *playout(
         typename Types::State &state,
         typename Types::Model &model,
         MatrixNode<Algorithm> *matrix_node)
@@ -117,7 +67,7 @@ private:
         {
             if (matrix_node->is_expanded)
             {
-                typename Types::Outcome outcome;
+                Outcome outcome;
 
                 this->select(matrix_node, outcome);
 
@@ -128,7 +78,7 @@ private:
                 ChanceNode<Algorithm> *chance_node = matrix_node->access(outcome.row_idx, outcome.col_idx);
                 MatrixNode<Algorithm> *matrix_node_next = chance_node->access(state.transition);
 
-                MatrixNode<Algorithm> *matrix_node_leaf = _playout(state, model, matrix_node_next);
+                MatrixNode<Algorithm> *matrix_node_leaf = playout(state, model, matrix_node_next);
 
                 outcome.row_value = matrix_node_leaf->inference.row_value;
                 outcome.col_value = matrix_node_leaf->inference.col_value;
@@ -147,4 +97,197 @@ private:
             return matrix_node;
         }
     }
+
+    void select(
+        MatrixNode<Algorithm> *matrix_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_select(
+            matrix_node,
+            outcome);
+    }
+
+    void init_stats(
+        int playouts,
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> *root)
+    {
+        static_cast<Algorithm *>(this)->_init_stats(
+            playouts,
+            state,
+            model,
+            root);
+    }
+    void expand(
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> *matrix_node)
+    {
+        static_cast<Algorithm *>(this)->_expand(
+            state,
+            model,
+            matrix_node);
+    }
+    void update_matrix_node(
+        MatrixNode<Algorithm> *matrix_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_update_matrix_node(
+            matrix_node,
+            outcome);
+    }
+
+    void update_chance_node(
+        ChanceNode<Algorithm> *chance_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_update_chance_node(
+            chance_node,
+            outcome);
+    }
+
+};
+
+/*
+TreeBandit with mutex's in each matrix node
+*/
+
+template <class Model, class Algorithm>
+class TreeBanditThreaded : public AbstractAlgorithm<Model>
+{
+public:
+    struct Outcome;
+    struct MatrixStats;
+    struct Types : AbstractAlgorithm<Model>::Types
+    {
+        using Outcome = TreeBanditThreaded::Outcome;
+        using MatrixNode = TreeBanditThreaded::MatrixStats;
+    };
+
+    struct Outcome
+    {
+        int row_idx, col_idx;
+        typename Types::Real row_value, col_value;
+        typename Types::Real row_mu, col_mu; // Actor policy at row_idx, col_idx.
+    };
+
+    struct MatrixStats
+    {
+        std::mutex mtx;
+    };
+
+    void run(
+        int playouts,
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> &matrix_node)
+    {
+        init_stats(playouts, state, model, &matrix_node);
+        for (int playout = 0; playout < playouts; ++playout)
+        {
+            this->playout(state, model, &matrix_node);
+        }
+    }
+
+private:
+
+    MatrixNode<Algorithm> *playout(
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> *matrix_node)
+    {
+
+        std::mutex &mtx = matrix_node->stats.mtx;
+
+        if (!matrix_node->is_terminal)
+        {
+            if (matrix_node->is_expanded)
+            {
+                Outcome outcome;
+
+                mtx.lock();
+                this->select(matrix_node, outcome);
+                mtx.unlock();
+
+                typename Types::Action row_action = matrix_node->actions.row_actions[outcome.row_idx];
+                typename Types::Action col_action = matrix_node->actions.col_actions[outcome.col_idx];
+                state.apply_actions(row_action, col_action);
+
+                ChanceNode<Algorithm> *chance_node = matrix_node->access(outcome.row_idx, outcome.col_idx);
+                MatrixNode<Algorithm> *matrix_node_next = chance_node->access(state.transition);
+
+                MatrixNode<Algorithm> *matrix_node_leaf = playout(state, model, matrix_node_next);
+
+                outcome.row_value = matrix_node_leaf->inference.row_value;
+                outcome.col_value = matrix_node_leaf->inference.col_value;
+                mtx.lock();
+                this->update_matrix_node(matrix_node, outcome);
+                this->update_chance_node(chance_node, outcome);
+                mtx.unlock();
+                return matrix_node_leaf;
+            }
+            else
+            {
+                mtx.lock();
+                this->expand(state, model, matrix_node);
+                mtx.unlock();
+                return matrix_node;
+            }
+        }
+        else
+        {
+            return matrix_node;
+        }
+    }
+
+    void select(
+        MatrixNode<Algorithm> *matrix_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_select(
+            matrix_node,
+            outcome);
+    }
+
+    void init_stats(
+        int playouts,
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> *root)
+    {
+        static_cast<Algorithm *>(this)->_init_stats(
+            playouts,
+            state,
+            model,
+            root);
+    }
+    void expand(
+        typename Types::State &state,
+        typename Types::Model &model,
+        MatrixNode<Algorithm> *matrix_node)
+    {
+        static_cast<Algorithm *>(this)->_expand(
+            state,
+            model,
+            matrix_node);
+    }
+    void update_matrix_node(
+        MatrixNode<Algorithm> *matrix_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_update_matrix_node(
+            matrix_node,
+            outcome);
+    }
+
+    void update_chance_node(
+        ChanceNode<Algorithm> *chance_node,
+        Outcome &outcome)
+    {
+        static_cast<Algorithm *>(this)->_update_chance_node(
+            chance_node,
+            outcome);
+    }
+
 };
